+++ ArribaSim/Scene/RegionLoader/Basic/RegionLoader.cs
@@ -97,7 +97,7 @@ namespace ArribaSim.Scene.RegionLoader.Basic
                     r.ServerURI = string.Format("http://{0}:{1}/", m_ExternalHostName, m_HttpPort);
                     r.Size.X = ((uint)regionEntry.GetInt("SizeX", 256) + 255) & (~(uint)255);
                     r.Size.Y = ((uint)regionEntry.GetInt("SizeY", 256) + 255) & (~(uint)255);
-                    r.Flags = (uint)RegionFlags.RegionOnline;
+                    r.Flags = RegionFlags.RegionOnline;
                     r.Owner.ID = regionEntry.GetString("OwnerID");
                     r.ScopeID = regionEntry.GetString("ScopeID", "00000000-0000-0000-0000-000000000000");
                     r.ServerHttpPort = m_HttpPort;
